requirements.txt
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
plotly==5.1.0
python-binance==0.7.9
python-dotenv==0.18.0
yfinance==0.1.63
matplotlib-inline==0.1.2           
websocket-client==1.1.0            
websockets==9.1                    